# PERF-#4727: Improve perf of `concat` operation