--- a/modin/core/dataframe/pandas/dataframe/dataframe.py
+++ b/modin/core/dataframe/pandas/dataframe/dataframe.py
@@ -2582,6 +2582,17 @@ class PandasDataframe(ClassLogger):
         """
         axis = Axis(axis)
         new_widths = None
+
+        def compute_new_width():
+            new_width = None
+            if self._column_widths_cache is not None and all(
+                o._column_widths_cache is not None for o in others
+            ):
+                new_width = self._column_widths_cache + [
+                    width for o in others for width in o._column_widths_cache
+                ]
+            return new_width
+
         # Fast path for equivalent columns and partitioning
         if (
             axis == Axis.ROW_WISE
@@ -2601,11 +2612,8 @@ class PandasDataframe(ClassLogger):
             left_parts = self._partitions
             right_parts = [o._partitions for o in others]
             new_lengths = self._row_lengths_cache
-            other_widths = [width for o in others for width in o._column_widths_cache]
-            if self._column_widths_cache is not None and all(
-                width is not None for width in other_widths
-            ):
-                new_widths = self._column_widths_cache + other_widths
+            # we can only do this for COL_WISE because `concat` might rebalance partitions for ROW_WISE
+            new_widths = compute_new_width()
         else:
             (
                 left_parts,
@@ -2617,6 +2625,7 @@ class PandasDataframe(ClassLogger):
             )
             if axis == Axis.COL_WISE:
                 new_lengths = partition_sizes_along_axis
+                new_widths = compute_new_width()
             else:
                 new_widths = partition_sizes_along_axis
         new_partitions = self._partition_mgr_cls.concat(