@anmyachev 2022-07-28
> @YarShev it would be nice for some benchmark results to confirm that things are working as expected

Results can be viewed [here](https://github.com/modin-project/modin/pull/4728#issue-1320042133).