@prutskov 2022-08-03
@anmyachev, why benchmark should be run without benchmark mode? In this case we will measure time of async binary opearations (for master branch). I got the next results with benchmark mode for `shape=10mx100`:

| Pandas | Modin master | The PR |
|-|:-:|-|
| 1.66 sec | 0.002 sec | 0.001 sec|

The PR adds async execution for `concat` operation, should this be as main message of this PR?